requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
google-generativeai==0.7.2
pandas==2.2.2
openpyxl==3.1.5
requests==2.31.0
beautifulsoup4==4.12.2
